Fix -Wshorten-64-to-32 warnings
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38932 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/diskimage/fsimage-probe.c
@@ -679,7 +679,7 @@ static int disk_image_check_for_dhd(disk_image_t *image)
     fsimage = image->media.fsimage;
     image->tracks = 65535;
 
-    blk = util_file_length(fsimage->fd);
+    blk = (unsigned int)util_file_length(fsimage->fd);
 
     /* first make sure the file is a multiple of 512 bytes and less than
        73728 bytes (which is the smallest possible DHD image */
